+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @0xsequence/hooks
 
+## 6.0.0-beta.4
+
+### Patch Changes
+
+- Ecosystem wallet config update
+
+## 6.0.0-beta.3
+
+### Patch Changes
+
+- ethauth fixes
+
+## 6.0.0-beta.2
+
+### Patch Changes
+
+- ethauth support
+
+## 6.0.0-beta.1
+
+### Patch Changes
+
+- 6.0.0 beta fixes
+
+## 6.0.0-beta.0
+
+### Major Changes
+
+- Connect package improvements for v6
+
 ## 5.4.8
 
 ### Patch Changes